+++ src/ZiggyServiceProvider.php
@@ -25,7 +25,9 @@ class ZiggyServiceProvider extends ServiceProvider
     protected function registerDirective(BladeCompiler $bladeCompiler)
     {
         $bladeCompiler->directive('routes', function ($group) {
-            return "<?php echo app('" . BladeRouteGenerator::class . "')->generate({$group}); ?>";
+            return app()->isLocal()
+                ? "<?php echo app('" . BladeRouteGenerator::class . "')->generate({$group}); ?>"
+                : app(BladeRouteGenerator::class)->generate($group);
         });
     }
 }