+++ tests/Unit/CommandRouteGeneratorTest.php
@@ -46,14 +46,7 @@ class CommandRouteGeneratorTest extends TestCase
 
         Artisan::call('ziggy:generate');
 
-        if ($this->laravelVersion(7)) {
-            $this->assertFileEquals('./tests/fixtures/ziggy.js', base_path('resources/js/ziggy.js'));
-        } else {
-            $this->assertSame(
-                str_replace(',"bindings":[]', '', file_get_contents(__DIR__ . '/../fixtures/ziggy.js')),
-                file_get_contents(base_path('resources/js/ziggy.js'))
-            );
-        }
+        $this->assertFileEquals('./tests/fixtures/ziggy.js', base_path('resources/js/ziggy.js'));
     }
 
     /** @test */
@@ -65,14 +58,7 @@ class CommandRouteGeneratorTest extends TestCase
 
         Artisan::call('ziggy:generate', ['--url' => 'http://example.org']);
 
-        if ($this->laravelVersion(7)) {
-            $this->assertFileEquals('./tests/fixtures/custom-url.js', base_path('resources/js/ziggy.js'));
-        } else {
-            $this->assertSame(
-                str_replace(',"bindings":[]', '', file_get_contents(__DIR__ . '/../fixtures/custom-url.js')),
-                file_get_contents(base_path('resources/js/ziggy.js'))
-            );
-        }
+        $this->assertFileEquals('./tests/fixtures/custom-url.js', base_path('resources/js/ziggy.js'));
     }
 
     /** @test */
@@ -88,14 +74,7 @@ class CommandRouteGeneratorTest extends TestCase
 
         Artisan::call('ziggy:generate');
 
-        if ($this->laravelVersion(7)) {
-            $this->assertFileEquals('./tests/fixtures/ziggy.js', base_path('resources/js/ziggy.js'));
-        } else {
-            $this->assertSame(
-                str_replace(',"bindings":[]', '', file_get_contents(__DIR__ . '/../fixtures/ziggy.js')),
-                file_get_contents(base_path('resources/js/ziggy.js'))
-            );
-        }
+        $this->assertFileEquals('./tests/fixtures/ziggy.js', base_path('resources/js/ziggy.js'));
     }
 
     /** @test */
@@ -114,13 +93,6 @@ class CommandRouteGeneratorTest extends TestCase
 
         Artisan::call('ziggy:generate', ['path' => 'resources/js/admin.js', '--group' => 'admin']);
 
-        if ($this->laravelVersion(7)) {
-            $this->assertFileEquals('./tests/fixtures/admin.js', base_path('resources/js/admin.js'));
-        } else {
-            $this->assertSame(
-                str_replace(',"bindings":[]', '', file_get_contents(__DIR__ . '/../fixtures/admin.js')),
-                file_get_contents(base_path('resources/js/admin.js'))
-            );
-        }
+        $this->assertFileEquals('./tests/fixtures/admin.js', base_path('resources/js/admin.js'));
     }
 }